The cylindrical get functions have been implemented in GriddedData, but they have not been rigorously tested yet. The analytic rankine model is functional again, now with additional wave numbers in VT.

+++ trunk/Radar/Ray.cpp
@@ -70,11 +70,11 @@ void Ray::setFirst_vel_gate(const int &value) {
   first_vel_gate = value;
 }
 
-void Ray::setRef_gatesp(const int &value) {
+void Ray::setRef_gatesp(const float &value) {
   ref_gatesp = value;
 }
 
-void Ray::setVel_gatesp(const int &value) {
+void Ray::setVel_gatesp(const float &value) {
   vel_gatesp = value;
 }
